requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
mysql-connector-python==9.5.0
pydantic==2.5.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6